+++ covenant-signer/keystore/cosmos/cosmoskeyretriever.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"strings"
+	"sync"
 
 	"github.com/babylonlabs-io/covenant-emulator/covenant-signer/config"
 	"github.com/babylonlabs-io/covenant-emulator/covenant-signer/signerapp"
@@ -13,8 +14,9 @@ import (
 var _ signerapp.PrivKeyRetriever = &CosmosKeyringRetriever{}
 
 type CosmosKeyringRetriever struct {
-	Kr         *ChainKeyringController
-	passphrase string
+	Kr           *ChainKeyringController
+	mu           sync.Mutex
+	btcecPrivKey *btcec.PrivateKey
 }
 
 func NewCosmosKeyringRetriever(cfg *config.CosmosKeyStoreConfig) (*CosmosKeyringRetriever, error) {
@@ -29,18 +31,55 @@ func NewCosmosKeyringRetriever(cfg *config.CosmosKeyStoreConfig) (*CosmosKeyring
 		return nil, err
 	}
 	return &CosmosKeyringRetriever{
-		Kr:         kc,
-		passphrase: cfg.Passphrase,
+		Kr:           kc,
+		btcecPrivKey: nil,
 	}, nil
 }
 
 func (k *CosmosKeyringRetriever) PrivKey(ctx context.Context) (*btcec.PrivateKey, error) {
-	privKey, err := k.Kr.GetChainPrivKey(k.passphrase)
+	k.mu.Lock()
+	defer k.mu.Unlock()
+
+	if k.btcecPrivKey == nil {
+		return nil, fmt.Errorf("private key is not unlocked. Please call Unlock() first")
+	}
+
+	return k.btcecPrivKey, nil
+}
+
+func (k *CosmosKeyringRetriever) Unlock(ctx context.Context, passphrase string) error {
+	k.mu.Lock()
+	defer k.mu.Unlock()
+
+	if k.btcecPrivKey != nil {
+		// already unlocked
+		return nil
+	}
+
+	privKey, err := k.Kr.GetChainPrivKey(passphrase)
 	if err != nil {
-		return nil, err
+		return fmt.Errorf("failed to unlock the key ring: %w", err)
 	}
 
 	btcecPrivKey, _ := btcec.PrivKeyFromBytes(privKey.Key)
 
-	return btcecPrivKey, nil
+	k.btcecPrivKey = btcecPrivKey
+	return nil
+}
+
+func (k *CosmosKeyringRetriever) Lock(ctx context.Context) error {
+	k.mu.Lock()
+	defer k.mu.Unlock()
+
+	if k.btcecPrivKey == nil {
+		// already locked
+		return nil
+	}
+
+	// First zero out the memory associated with the private key
+	k.btcecPrivKey.Zero()
+	// Clear the reference to the private key
+	k.btcecPrivKey = nil
+
+	return nil
 }