+++ covenant-signer/signerservice/types/publickey.go
@@ -4,3 +4,4 @@ type GetPublicKeyResponse struct {
 	// hex encoded 33 byte public key
 	PublicKey string `json:"public_key_hex"`
 }
+